+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - no unreleased changes so far
+## [0.6.0] - 2021-06-11
+### Added
+- New attribute coUsers for users in secondary role
+
+### Changed
+- Enum used for role and enrollmentStatus attributes
+
 ## [0.5.2] - 2021-06-10
 ### Fixed
 - Charging settings setting and request status response
@@ -67,7 +74,8 @@ Minor fix in observer interface
 ## [0.1.0] - 2021-05-26
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.5.2...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.6.0
 [0.5.2]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.5.2
 [0.5.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.5.1
 [0.5.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.5.0